Move csi under /cmd
