+++ .github/workflows/csi-test.yml
@@ -13,7 +13,8 @@ on:
       - "docs/**"
   pull_request:
     paths:
-      - "clients/csi/**"
+      - "cmd/csi/**"
+      - "internal/csi/**"
       - "internal/server/openapi/api_model_registry_service*"
       - "pkg/openapi/**"
       # csi build depends on base go.mod https://github.com/kubeflow/model-registry/issues/311
