+++ cmd/csi/samples/modelregistry.clusterstoragecontainer.yaml
